@@ -2,12 +2,25 @@
 
 namespace App\Exceptions;
 
+use App\Traits\RespondsWithHttpStatus;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\Exceptions\PostTooLargeException;
+use Illuminate\Http\Exceptions\ThrottleRequestsException;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Response;
+use Illuminate\Validation\ValidationException;
+use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Throwable;
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
 
 class Handler extends ExceptionHandler
 {
+    use RespondsWithHttpStatus;
+
     /**
      * A list of the exception types that are not reported.
      *
@@ -33,18 +46,45 @@ class Handler extends ExceptionHandler
      *
      * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->reportable(function (Throwable $e) {
             //
         });
     }
 
-    protected function unauthenticated($request, AuthenticationException $exception)
+    public function render($request, Throwable $e): Response|JsonResponse|\Symfony\Component\HttpFoundation\Response
+    {
+        if ($request->expectsJson()) {
+            if ($e instanceof AuthorizationException)
+                return $this->failure(message: $e->getMessage() ?? 'You don\'t own this resource!', status: 403);
+
+            if ($e instanceof NotFoundHttpException)
+                return $this->failure(message: 'Invalid route!', status: 404);
+
+            if ($e instanceof ModelNotFoundException)
+                return $this->failure(message: 'Model not found!', status: 404);
+
+            if ($e instanceof MethodNotAllowedHttpException)
+                return $this->failure(details:  $e->getMessage(), status: 405);
+
+            if ($e instanceof PostTooLargeException)
+                return $this->failure(message: 'Payload too large!', status: 413);
+
+            if ($e instanceof ThrottleRequestsException)
+                return $this->failure(message: 'Too many requests!', status: 429);
+
+            if ($e instanceof ValidationException)
+                return $this->failure(message: 'Invalid payload!', details: $e->errors());
+        }
+        return parent::render($request, $e);
+    }
+
+    protected function unauthenticated($request, AuthenticationException $exception): JsonResponse|\Symfony\Component\HttpFoundation\Response|RedirectResponse
     {
 
         if ($request->expectsJson()) {
-            return response()->json(['error' => 'Unauthenticated.'], 401);
+            return $this->failure(message: 'Unauthenticated.', status: 401);
         }
 
         if ($request->is('admin/*')) {
