@@ -184,8 +184,8 @@ class InvestmentController extends Controller
             }
             $datum['package'] = '<span class="text-gray-600 fw-bolder d-block fs-6">'.$investment->package['name'].'</span>';
             $datum['slots'] = '<span class="text-gray-600 fw-bolder d-block fs-6">'.$investment['slots'].'</span>';
-            $datum['total_invested'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency() .' '.number_format($investment['amount']).'</span>';
-            $datum['expected_returns'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency() .' '.number_format($investment['total_return']).'</span>';
+            $datum['total_invested'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency() .' '.number_format($investment['amount'], 2).'</span>';
+            $datum['expected_returns'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency() .' '.number_format($investment['total_return'], 2).'</span>';
             $datum['investment_date'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">'.$investment->investment_date->format('M d, Y').'</span>';
             $datum['status'] = $status;
             $datum['payment'] = $payment;
@@ -229,7 +229,7 @@ class InvestmentController extends Controller
                 ]);
                 if ($transaction) {
                     $referee->wallet()->increment('balance', 50);
-                    $referral->update(['paid' => true]);
+                    $referral->update(['paid' => true, 'amount' => 50]);
                     try {
                         NotificationController::sendReferralTransactionNotification($transaction);
                     } catch (Exception $e) {
