@@ -58,11 +58,11 @@ class PackageController extends Controller
             'roi' => ['required', 'numeric'],
             'start_date' => ['required', 'date'],
             'slots' => ['required_if:type,farm'],
-            'duration' => ['required_if:type,farm', 'numeric'],
             'price' => ['required', 'numeric', 'gt:0'],
             'milestones' => ['required_if:type,plant,tractor'],
             'duration' => ['required_if:type,farm', 'numeric'],
             'payout_mode' => ['required_if:type,plant,tractor'],
+            'months' => ['required_if:payout_mode,custom'],
             'rollover' => ['sometimes'],
             'description' => ['required'],
             'image' => ['required_if:type,plant,tractor', 'mimes:jpeg,jpg,png', 'max:1024'],
@@ -99,11 +99,11 @@ class PackageController extends Controller
             'roi' => ['required', 'numeric'],
             'start_date' => ['required', 'date'],
             'slots' => ['required_if:type,farm'],
-            'duration' => ['required_if:type,farm', 'numeric'],
             'price' => ['required', 'numeric', 'gt:0'],
             'milestones' => ['required_if:type,plant,tractor'],
             'duration' => ['required_if:type,farm', 'numeric'],
             'payout_mode' => ['required_if:type,plant,tractor'],
+            'months' => ['required_if:payout_mode,custom'],
             'rollover' => ['sometimes'],
             'description' => ['required'],
             'image' => ['mimes:jpeg,jpg,png', 'max:1024'],
