@@ -41,7 +41,7 @@ class TransactionController extends Controller
         $user->wallet()->increment('balance', $request['amount']);
         $transaction = $user->transactions()->create([
             'type' => 'deposit', 'amount' => $request['amount'],
-            'amount_in_naira' => \App\Http\Controllers\OnlinePaymentController::getAmountInNaira($request['amount']),
+            'amount_in_naira' => \App\Http\Controllers\OnlinePaymentController::getAmountInNaira($request['amount'], $user),
             'description' => 'Deposit by '.env('APP_NAME'),
             'method' => 'deposit' ,'status' => 'approved'
         ]);
@@ -78,7 +78,7 @@ class TransactionController extends Controller
         $user->wallet()->decrement('balance', $request['amount']);
         $transaction = $user->transactions()->create([
             'type' => 'withdrawal', 'amount' => $request['amount'], 'method' => 'wallet',
-            'amount_in_naira' => \App\Http\Controllers\OnlinePaymentController::getAmountInNaira($request['amount']),
+            'amount_in_naira' => \App\Http\Controllers\OnlinePaymentController::getAmountInNaira($request['amount'], $user),
             'description' => 'Withdrawal by '.env('APP_NAME'), 'status' => 'approved'
         ]);
         if ($transaction) {
@@ -296,8 +296,8 @@ class TransactionController extends Controller
             }else{
                 $datum['name'] = ucwords($transaction->user['name']);
             }
-            $datum['amount'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency() .' '.number_format((int)$transaction['amount']).'</span>';
-            $datum['amount_in_naira'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;"> ₦ '.number_format((int)$transaction['amount_in_naira']).'</span>';
+            $datum['amount'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">' . getCurrency($transaction->user) .' '.number_format($transaction['amount'], 2).'</span>';
+            $datum['amount_in_naira'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;"> ₦ '.number_format($transaction['amount_in_naira'], 2).'</span>';
             $datum['description'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">'.$transaction['description'].'</span>';
             $datum['date'] = '<span class="text-gray-600 fw-bolder d-block fs-6" style="white-space: nowrap;">'.$transaction['created_at']->format('M d, Y \a\t h:i A').'</span>';
             $datum['details'] = '<span class="text-gray-600 fw-bolder d-block fs-6 text-center" style="white-space: nowrap;">'.$details.'</span>';
