@@ -5,90 +5,51 @@ namespace App\Policies;
 use App\Models\Document;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Illuminate\Auth\Access\Response;
 
 class DocumentPolicy
 {
     use HandlesAuthorization;
 
     /**
-     * Determine whether the user can view any models.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Auth\Access\Response|bool
-     */
-    public function viewAny(User $user)
-    {
-        //
-    }
-
-    /**
      * Determine whether the user can view the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Document  $document
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Document $document
+     * @return Response|bool
      */
-    public function view(User $user, Document $document)
+    public function view(User $user, Document $document): Response|bool
     {
-        //
-    }
-
-    /**
-     * Determine whether the user can create models.
-     *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Auth\Access\Response|bool
-     */
-    public function create(User $user)
-    {
-        //
+        return $user['id'] === $document['user_id']
+            ? Response::allow()
+            : Response::deny('You don\'t own this document!', 403);
     }
 
     /**
      * Determine whether the user can update the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Document  $document
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Document $document
+     * @return Response|bool
      */
-    public function update(User $user, Document $document)
+    public function update(User $user, Document $document): Response|bool
     {
-        //
+        return $user['id'] === $document['user_id']
+            ? Response::allow()
+            : Response::deny('You don\'t own this document!', 403);
     }
 
     /**
      * Determine whether the user can delete the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Document  $document
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Document $document
+     * @return Response|bool
      */
-    public function delete(User $user, Document $document)
+    public function delete(User $user, Document $document): Response|bool
     {
-        //
-    }
-
-    /**
-     * Determine whether the user can restore the model.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Document  $document
-     * @return \Illuminate\Auth\Access\Response|bool
-     */
-    public function restore(User $user, Document $document)
-    {
-        //
-    }
-
-    /**
-     * Determine whether the user can permanently delete the model.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Document  $document
-     * @return \Illuminate\Auth\Access\Response|bool
-     */
-    public function forceDelete(User $user, Document $document)
-    {
-        //
+        return $user['id'] === $document['user_id']
+            ? Response::allow()
+            : Response::deny('You don\'t own this document!', 403);
     }
 }
