@@ -5,6 +5,7 @@ namespace App\Policies;
 use App\Models\Investment;
 use App\Models\User;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Illuminate\Auth\Access\Response;
 
 class InvestmentPolicy
 {
@@ -13,8 +14,8 @@ class InvestmentPolicy
     /**
      * Determine whether the user can view any models.
      *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @return Response|bool
      */
     public function viewAny(User $user)
     {
@@ -24,20 +25,22 @@ class InvestmentPolicy
     /**
      * Determine whether the user can view the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Investment  $investment
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Investment $investment
+     * @return Response|bool
      */
-    public function view(User $user, Investment $investment)
+    public function view(User $user, Investment $investment): Response|bool
     {
-        //
+        return $user['id'] === $investment['user_id']
+            ? Response::allow()
+            : Response::deny('You don\'t own this investment!', 403);
     }
 
     /**
      * Determine whether the user can create models.
      *
-     * @param  \App\Models\User  $user
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @return Response|bool
      */
     public function create(User $user)
     {
@@ -47,9 +50,9 @@ class InvestmentPolicy
     /**
      * Determine whether the user can update the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Investment  $investment
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Investment $investment
+     * @return Response|bool
      */
     public function update(User $user, Investment $investment)
     {
@@ -59,9 +62,9 @@ class InvestmentPolicy
     /**
      * Determine whether the user can delete the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Investment  $investment
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Investment $investment
+     * @return Response|bool
      */
     public function delete(User $user, Investment $investment)
     {
@@ -71,9 +74,9 @@ class InvestmentPolicy
     /**
      * Determine whether the user can restore the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Investment  $investment
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Investment $investment
+     * @return Response|bool
      */
     public function restore(User $user, Investment $investment)
     {
@@ -83,9 +86,9 @@ class InvestmentPolicy
     /**
      * Determine whether the user can permanently delete the model.
      *
-     * @param  \App\Models\User  $user
-     * @param  \App\Models\Investment  $investment
-     * @return \Illuminate\Auth\Access\Response|bool
+     * @param User $user
+     * @param Investment $investment
+     * @return Response|bool
      */
     public function forceDelete(User $user, Investment $investment)
     {
