@@ -16,7 +16,7 @@ class PackageFactory extends Factory
         return [
             'name' => ucwords($this->faker->words()),
             'duration' => $this->faker->numberBetween(1,10),
-            'payout_mode' => $this->faker->randomElement(['single', 'monthly', 'quarterly', 'semi-annually', 'biannually', 'annually' ]),
+            'payout_mode' => $this->faker->randomElement(['single', 'monthly', 'quarterly', 'semi-annually', 'biannually', 'annually', 'custom']),
             "duration_mode" => $this->faker->randomElement(['day', 'month', 'year']),
             'roi' => $this->faker->numberBetween(10, 100),
             'type' => $this->faker->randomElement(['plant', 'farm']),
