@@ -188,7 +188,7 @@
                             <td class="ps-4"><span class="text-dark fw-bolder d-block mb-1 fs-6 text-nowrap">{{ $key + 1 }}</span></td>
                             <td><a href="{{ route('admin.packages.show', ['package' => $package['id'], 'type' => $package['type']]) }}"><span class="text-primary fw-bolder d-block fs-6 text-nowrap">{{ $package['name'] }}</span></a></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6 text-nowrap">{{ $package['roi'] }}%</span></td>
-                            <td><span class="text-gray-600 fw-bolder d-block fs-6 text-nowrap">{{ getCurrency() }} {{ number_format($package['price']) }}</span></td>
+                            <td><span class="text-gray-600 fw-bolder d-block fs-6 text-nowrap">{{ getCurrency() }} {{ number_format($package['price'], 2) }}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6 text-nowrap">{{ $package['start_date']->format('M d, Y \a\t h:i A') }}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6 text-nowrap">{{ ucwords($package['type'])}}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6 text-center">{{ $package->investments()->where('payment', 'approved')->count() }}</span></td>
