@@ -140,7 +140,7 @@
                     <tr class="fw-bolder text-muted bg-light">
                         <th class="ps-4 text-dark rounded-start">SN</th>
                         <th class="text-dark">Name</th>
-                        <th class="text-dark">Amount(USD)</th>
+                        <th class="text-dark">Amount</th>
                         <th class="text-dark">Amount(NGN)</th>
                         <th class="text-dark">Description</th>
                         <th class="text-dark">Date</th>
