@@ -7,7 +7,7 @@
 @endsection
 
 @section('breadCrumbs')
-    <li class="breadcrumb-item"><a href="javascript:void()" class="text-muted">Transactions</a></li>
+    <li class="breadcrumb-item"><a href="javascript:void(0)" class="text-muted">Transactions</a></li>
 @endsection
 
 @section('content')
@@ -175,7 +175,7 @@
                         @endphp
                         <tr>
                             <td class="ps-4"><span class="text-dark fw-bolder d-block mb-1 fs-6 text-nowrap">{{ $key + 1 }}</span></td>
-                            <td><span class="text-gray-600 fw-bolder d-block fs-6">{{ getCurrency() }} {{ number_format($transaction['amount']) }}</span></td>
+                            <td><span class="text-gray-600 fw-bolder d-block fs-6">{{ getCurrency() }} {{ number_format($transaction['amount'], 2) }}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6">{{ $transaction['description'] }}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6">{{ $transaction['created_at']->format('M d, Y') }}</span></td>
                             <td><span class="text-gray-600 fw-bolder d-block fs-6 text-center">{!! $details !!}</span></td>
